Refactor

+++ WebFramework/src/LoadBalancer/LoadBalancer.cpp
@@ -22,7 +22,7 @@ namespace framework
 
 			const string& loadBalancerIp = parser.get<string>(json_settings::loadBalancerIpKey);
 			const string& loadBalancerPort = parser.get<string>(json_settings::loadBalancerPortKey);
-			int64_t loadBalancerTimeout = parser.get<int64_t>(json_settings::loadBalancerTimeoutKey);
+			DWORD loadBalancerTimeout = static_cast<DWORD>(parser.get<int64_t>(json_settings::loadBalancerTimeoutKey));
 			const smartPointer<json::JSONParser::objectType>& listOfServers = parser.get<smartPointer<json::JSONParser::objectType>>(json_settings::listOfServersKey);
 			unordered_map<string, vector<string>> allServers;
